Add files via upload
First, I want to say about the capabilities of my bot and what it can do, what commands we can enter so that the bot works correctly. In short, I created this bot to help students like me. With this bot we can quickly find all the lessons on the site. The names of our teachers and how to contact them. You see, during an online class, many students are late for class. Because of this, we start the lesson 5 or 10 minutes late. I don’t blame them, because it’s not easy to find tutorials and links to zoom on the site. And it is still not easy to contact teachers. That's why I wanted to create such a bot.
 Press any word and the bot will enter to us:
I do not have such kind of command
If you get bored click on / entertainment
if you need help click on / help
As you can see, the bot will offer us two commands
Well, if you are tired and want to relax, then click on the first command / entertainment and it will offer us all sorts of interesting YouTube channels. For example football, news, animal world, movies or video bloggers. Well, this is mainly if you are already very tired or bored.
And if you want to immediately get to work, then click on the command / help.
There will appear 4 commands. For example, if you want to find lessons in the OCC, then click on the first command and it will show a list of all lessons in the OCS. There you will choose what you need, and there you will see a link to the lesson, the teacher's name and the teacher's mail.
In short, it is very easy to understand.
And other teams are like additional lessons. There are commands / python_language, / html_language and / creat_a_bot.
As you understand, 2 is a command for learning python, and 3 is a command for html. Press the 4-command if you want to create your own bot by yourself, I think this will make the work of many students easier. That's all for now. That's all i could do